+++ src/Fixer/NoUselessSprintfFixer.php
@@ -13,16 +13,25 @@ declare(strict_types=1);
 
 namespace PhpCsFixerCustomFixers\Fixer;
 
+use PhpCsFixer\Fixer\DeprecatedFixerInterface;
 use PhpCsFixer\FixerDefinition\CodeSample;
 use PhpCsFixer\FixerDefinition\FixerDefinition;
 use PhpCsFixer\FixerDefinition\FixerDefinitionInterface;
-use PhpCsFixer\Tokenizer\Analyzer\ArgumentsAnalyzer;
-use PhpCsFixer\Tokenizer\Analyzer\FunctionsAnalyzer;
-use PhpCsFixer\Tokenizer\Token;
 use PhpCsFixer\Tokenizer\Tokens;
 
-final class NoUselessSprintfFixer extends AbstractFixer
+/**
+ * @deprecated
+ */
+final class NoUselessSprintfFixer extends AbstractFixer implements DeprecatedFixerInterface
 {
+    /** @var \PhpCsFixer\Fixer\FunctionNotation\NoUselessSprintfFixer */
+    private $fixer;
+
+    public function __construct()
+    {
+        $this->fixer = new \PhpCsFixer\Fixer\FunctionNotation\NoUselessSprintfFixer();
+    }
+
     public function getDefinition(): FixerDefinitionInterface
     {
         return new FixerDefinition(
@@ -35,79 +44,29 @@ final class NoUselessSprintfFixer extends AbstractFixer
 
     public function getPriority(): int
     {
-        return 0;
+        return $this->fixer->getPriority();
     }
 
     public function isCandidate(Tokens $tokens): bool
     {
-        return $tokens->isTokenKindFound(T_STRING);
+        return $this->fixer->isCandidate($tokens);
     }
 
     public function isRisky(): bool
     {
-        return true;
+        return $this->fixer->isRisky();
     }
 
     public function fix(\SplFileInfo $file, Tokens $tokens): void
     {
-        $argumentsAnalyzer = new ArgumentsAnalyzer();
-        $functionsAnalyzer = new FunctionsAnalyzer();
-
-        for ($index = $tokens->count() - 1; $index > 0; $index--) {
-            /** @var Token $token */
-            $token = $tokens[$index];
-
-            if (!$token->equals([T_STRING, 'sprintf'], false)) {
-                continue;
-            }
-
-            if (!$functionsAnalyzer->isGlobalFunctionCall($tokens, $index)) {
-                continue;
-            }
-
-            /** @var int $openParenthesisIndex */
-            $openParenthesisIndex = $tokens->getNextTokenOfKind($index, ['(']);
-
-            $closeParenthesisIndex = $tokens->findBlockEnd(Tokens::BLOCK_TYPE_PARENTHESIS_BRACE, $openParenthesisIndex);
-
-            if ($argumentsAnalyzer->countArguments($tokens, $openParenthesisIndex, $closeParenthesisIndex) !== 1) {
-                continue;
-            }
-
-            /** @var int $afterOpenParenthesisIndex */
-            $afterOpenParenthesisIndex = $tokens->getNextMeaningfulToken($openParenthesisIndex);
-
-            /** @var Token $afterOpenParenthesisToken */
-            $afterOpenParenthesisToken = $tokens[$afterOpenParenthesisIndex];
-
-            if ($afterOpenParenthesisToken->isGivenKind(T_ELLIPSIS)) {
-                continue;
-            }
-
-            /** @var int $prevIndex */
-            $prevIndex = $tokens->getPrevMeaningfulToken($index);
-
-            /** @var Token $prevToken */
-            $prevToken = $tokens[$prevIndex];
-
-            if ($prevToken->isGivenKind(T_NS_SEPARATOR)) {
-                $this->removeTokenAndSiblingWhitespace($tokens, $prevIndex, 1);
-            }
-            $this->removeTokenAndSiblingWhitespace($tokens, $index, 1);
-            $this->removeTokenAndSiblingWhitespace($tokens, $openParenthesisIndex, 1);
-            $this->removeTokenAndSiblingWhitespace($tokens, $closeParenthesisIndex, -1);
-        }
+        $this->fixer->fix($file, $tokens);
     }
 
-    private function removeTokenAndSiblingWhitespace(Tokens $tokens, int $index, int $direction): void
+    /**
+     * @return string[]
+     */
+    public function getSuccessorsNames(): array
     {
-        $tokens->clearAt($index);
-
-        /** @var Token $siblingToken */
-        $siblingToken = $tokens[$index + $direction];
-
-        if ($siblingToken->isWhitespace()) {
-            $tokens->clearAt($index + $direction);
-        }
+        return [$this->fixer->getName()];
     }
 }